refactor: 122.11a -- consolidate the per-frame drains into one module
Group B shrank the god functions as intended, but `app_impl.rs` itself
grew from 4,878 to 5,856 lines because each extraction's doc comments
and tests landed in the same file. One coherent concept accounted for
most of it: per-frame pane-event draining -- everything that walks
`win.tabs` once per frame to drain a channel and stage the results.

Two sub-agents, asked independently during 122.7 and 122.8, proposed
the same home. It now exists:

  app_impl.rs    5,856 -> 4,990   (-866)
  frame_drain.rs           938    (new)

That puts `app_impl.rs` below where it stood before Group B started,
and below the 5,319 it was at when Task 122 was activated.

Moved: `drain_command_finished_events`, `process_dead_panes`,
`DeadPaneOutcome`, `WindowFocus`, `drain_window_manipulation_commands`,
`WindowManipulationEvents`, and the four test helpers, with their tests.

Deliberately NOT moved: `stage_frame_damage` and `stage_chrome_signals`
with their input/observation structs. Those are damage and signal
*staging*, not channel draining -- a different concept, and folding
them in would make the module a grab-bag rather than a home.

`route_window_manipulation_events` stayed a method on `FreminalGui`.
Converting it to a free function would have meant threading five
distinct pieces of `self` state through the signature, or passing
`&FreminalGui` and defeating the point. Only its inputs moved.

A pure move: no function body changed apart from import paths, no test
assertion changed, nothing renamed, and the test count is identical at
7,373 before and after.

One visibility widening, disclosed rather than buried:
`WindowFocus::is_focused` went from private to `pub(super)`, because
`route_window_manipulation_events` stayed in `app_impl.rs` and calls it
at two sites. Still within the `pub(super)` ceiling; nothing became
`pub`.

Verification: cargo test --all; cargo test --all --features
frame-profiling; cargo clippy --workspace --all-targets; cargo clippy
--all-targets --all-features -D warnings; cargo machete; cargo fmt
--check.
